chore: move DFA caching into separate files
